fixed actions.ts
