+++ app/[locale]/page.tsx
@@ -4,7 +4,7 @@ import Container from "@/app/[locale]/components/container";
 import { Navbar } from "@/app/[locale]/components/navbar";
 import Carrousel from "@/app/[locale]/components/carrousel";
 import { MorePosts } from "@/app/[locale]/components/more-posts";
-import { getPaginatedPosts } from "@/app/actions";
+import { getPaginatedPosts } from "@/app/lib/queries";
 
 type Locale = "en" | "es";
