+++ app/api/posts/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { getPaginatedPosts } from "@/app/actions";
+import { getPaginatedPosts } from "@/app/lib/queries";
 
 export async function GET(req: Request) {
   const { searchParams } = new URL(req.url);
